+++ src/truflation/data/connector/factory.py
@@ -1,0 +1,4 @@
+connector_factory_list = []
+
+def add_connector_factory(factory_function) -> None:
+    connector_factory_list.append(factory_function)
