+++ src/truflation/data/connector/gsheet.py
@@ -1,11 +1,14 @@
-
+import logging
 import gspread
 import pandas as pd
 
 from typing import Optional
 from gspread_pandas import Spread, Client
-from truflation.data.connectors.base import Connector
+from tenacity import retry, stop_after_attempt, wait_fixed, before_log
 
+from .base import Connector
+
+logger = logging.getLogger(__name__)
 
 class ConnectorGoogleSheets(Connector):
     """
@@ -34,9 +37,14 @@ class ConnectorGoogleSheets(Connector):
         self.client = self.my_client if self.path_root is not None \
             else None
 
+    @retry(
+        stop=stop_after_attempt(3),
+        wait=wait_fixed(2),
+        before=before_log(logger, logging.INFO)
+    )  # will retry 3 times with a 2-second wait between attempts.
     def read_all(self, sheet_id: str, *args, **kwargs) -> Optional[pd.DataFrame]:
         """
-        Read all data from a Google Sheets document.
+        Read all data from a Google Sheets document with retry logic.
 
         Args:
             sheet_id (str): The ID of the Google Sheets document.
@@ -55,7 +63,7 @@ class ConnectorGoogleSheets(Connector):
                 df.rename(columns={'Date': 'date'}, inplace=True)
                 return df
             except Exception as e:
-                self.logging_manager.log_error(f'Error fetching data from Google Sheets: {e}')
+                self.logging_manager.log_error(f'Error fetching data from Google Sheets URL: {url}. Error: {e}')
                 return None
         try:
             self.logging_manager.log_info('Fetching data from Google Sheets using gspread...')
@@ -81,10 +89,15 @@ class ConnectorGoogleSheets(Connector):
                     df[column] = df[column].astype(int)
             return df
 
-        except gspread.exceptions.SpreadsheetNotFound:
-            self.logging_manager.log_error('Google Sheets spreadsheet not found.')
+        except gspread.exceptions.SpreadsheetNotFound as e:
+            self.logging_manager.log_error(f'Google Sheets spreadsheet not found for ID: {sheet_id}. Error: {e}')
             return None
-
+        except gspread.exceptions.APIError as e:
+            self.logging_manager.log_error(f'API error while accessing Google Sheets with ID: {sheet_id}. Error: {e}')
+            raise
+        except Exception as e:
+            self.logging_manager.log_error(f'Unexpected error occurred while reading Google Sheets: {e}')
+            raise
     def write_all(self, df, *args, **kwargs):
         key = kwargs.get('key', self.default_key)
         spread = Spread(key, create_spread=True)
@@ -96,7 +109,7 @@ class ConnectorGoogleSheets(Connector):
         else:
             dims = spread.get_sheet_dims()
             self.logging_manager.log_debug(f'Sheet dimensions: {dims}')
-            self.logger.debug(dims)
+            logger.debug(dims)
             start_row = dims[0] + 1 if dims[0] > 1 else 1
             self.logging_manager.log_info(f'Appending data to sheet starting from row {start_row}.')
             spread.df_to_sheet(df.astype(str), start=(start_row, 1), headers=dims[0] < 2)
