+++ src/truflation/data/connector/json.py
@@ -3,7 +3,7 @@ import os
 import pandas as pd
 from pathlib import Path
 from typing import Any
-from truflation.data.connectors.base import Connector
+from .base import Connector
 
 
 class ConnectorJson(Connector):
