+++ src/truflation/data/connector/kwil.py
@@ -17,8 +17,9 @@ import aiohttp
 import logging
 
 import truflation.data
-from truflation.data.connector import add_connector_factory
-from truflation.data.connectors.base import Connector
+from .factory import add_connector_factory
+from .base import Connector
+
 from eth_utils import to_checksum_address
 
 from dotenv import load_dotenv
