+++ tests/test_kwil.py
@@ -1,8 +1,7 @@
 import os
 import unittest
 
-import truflation.data.connectors.kwil as kwil
-from truflation.data.connector import connector_factory
+from truflation.data.connector import connector_factory, ConnectorKwil
 
 @unittest.skipIf('KWIL_USER' not in os.environ, "skipping test")
 class TestSimple(unittest.TestCase):
@@ -13,12 +12,12 @@ class TestSimple(unittest.TestCase):
         self.assertTrue(output.strip() == 'pong')
 
     def test_simple(self):
-        connector = kwil.ConnectorKwil()
+        connector = ConnectorKwil()
         output = connector.ping()
         self.assertTrue(output.strip() == 'pong')
         version = connector.version()
 
     def test_list_databases(self):
-        connector = kwil.ConnectorKwil()
+        connector = ConnectorKwil()
         output = connector.list_databases()
